Made int to char comparison explicit so keyboard input works properly on 64-bit systems.
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@591 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/samples/c/facedetect.c
@@ -134,7 +134,7 @@ int main( int argc, char** argv )
 
 void detect_and_draw( IplImage* img )
 {
-    int scale = 1.3;
+    double scale = 1.3;
     IplImage* gray = cvCreateImage( cvSize(img->width,img->height), 8, 1 );
     IplImage* small_img = cvCreateImage( cvSize(img->width/scale,img->height/scale), 8, 1 );
     CvPoint pt1, pt2;
